test: use tap instead of nodeunit
