+++ package.json
@@ -16,12 +16,11 @@
   },
   "devDependencies": {
     "mkdirp": "^0.5.1",
-    "multicouch": "^0.8.2",
-    "nodeunit": "^0.9.1",
     "request": "^2.60.0",
     "rimraf": "^2.4.2",
     "snazzy": "^1.0.3",
-    "standard": "^4.5.4"
+    "standard": "^4.5.4",
+    "tap": "^1.3.2"
   },
   "homepage": "https://github.com/hoodiehq/hoodie-plugins-manager",
   "keywords": [
@@ -37,6 +36,6 @@
     "url": "https://github.com/hoodiehq/hoodie-plugins-manager.git"
   },
   "scripts": {
-    "test": "snazzy"
+    "test": "snazzy && tap test/*.js"
   }
 }
