+++ test/fixtures/couch.ini
@@ -1,0 +1,14 @@
+[couchdb]
+database_dir = ./data
+view_index_dir = ./data
+uri_file = ./data/couch.uri
+
+[httpd]
+port = 5984
+authentication_handlers = {couch_httpd_oauth, oauth_authentication_handler},{couch_httpd_auth, default_authentication_handler},{couch_httpd_auth, cookie_authentication_handler}
+
+[log]
+file = ./data/couch.log
+
+[admins]
+admin = -pbkdf2-74deb2b6e034354ff9998876eb4ad41764c91a8e,6000fe1ae25935b6b1cb385d502a1c37,10
